pyproject.toml
[project]
name = "zaphiro-tech-test"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.128.0",
    "polars[adbc,connectorx]>=1.37.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.31",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[dependency-groups]
dev = [
    "pyrefly>=0.49.0",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]

[tool.pytest]
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]
